SK-1664 move ColumnValue class to tokens package
