+++ v2/src/main/java/com/skyflow/vault/tokens/TokenizeRequest.java
@@ -1,7 +1,5 @@
 package com.skyflow.vault.tokens;
 
-import com.skyflow.utils.ColumnValue;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,11 +21,11 @@ public class TokenizeRequest {
     public static final class TokenizeRequestBuilder {
         private List<ColumnValue> columnValues;
 
-        public TokenizeRequestBuilder() {
+        private TokenizeRequestBuilder() {
             this.columnValues = new ArrayList<>();
         }
 
-        public TokenizeRequestBuilder columnValues(List<ColumnValue> columnValues) {
+        public TokenizeRequestBuilder values(List<ColumnValue> columnValues) {
             this.columnValues = columnValues;
             return this;
         }
